Rename plugins to remove the trailing 's'
