+++ tensorboard/BUILD
@@ -28,12 +28,12 @@ py_binary(
         "//tensorboard/backend/event_processing:event_file_inspector",
         "//tensorboard/plugins/audio:audio_plugin",
         "//tensorboard/plugins/core:core_plugin",
-        "//tensorboard/plugins/distributions:distributions_plugin",
-        "//tensorboard/plugins/graphs:graphs_plugin",
-        "//tensorboard/plugins/histograms:histograms_plugin",
-        "//tensorboard/plugins/images:images_plugin",
+        "//tensorboard/plugins/distribution:distributions_plugin",
+        "//tensorboard/plugins/graph:graphs_plugin",
+        "//tensorboard/plugins/histogram:histograms_plugin",
+        "//tensorboard/plugins/image:images_plugin",
         "//tensorboard/plugins/projector:projector_plugin",
-        "//tensorboard/plugins/scalars:scalars_plugin",
+        "//tensorboard/plugins/scalar:scalars_plugin",
         "//tensorboard/plugins/text:text_plugin",
         "@org_pocoo_werkzeug//:werkzeug",
     ],
