+++ tensorboard/backend/event_processing/BUILD
@@ -83,7 +83,7 @@ py_library(
         ":plugin_asset_util",
         ":reservoir",
         "//tensorboard:expect_tensorflow_installed",
-        "//tensorboard/plugins/distributions:compressor",
+        "//tensorboard/plugins/distribution:compressor",
     ],
 )
 
@@ -95,7 +95,7 @@ py_test(
     deps = [
         ":event_accumulator",
         "//tensorboard:expect_tensorflow_installed",
-        "//tensorboard/plugins/distributions:compressor",
+        "//tensorboard/plugins/distribution:compressor",
     ],
 )
