+++ tensorboard/backend/event_processing/event_accumulator.py
@@ -27,7 +27,7 @@ from tensorboard.backend.event_processing import directory_watcher
 from tensorboard.backend.event_processing import event_file_loader
 from tensorboard.backend.event_processing import plugin_asset_util
 from tensorboard.backend.event_processing import reservoir
-from tensorboard.plugins.distributions import compressor
+from tensorboard.plugins.distribution import compressor
 
 namedtuple = collections.namedtuple
 ScalarEvent = namedtuple('ScalarEvent', ['wall_time', 'step', 'value'])
