+++ tensorboard/components/tf_tensorboard/BUILD
@@ -15,20 +15,20 @@ ts_web_library(
     path = "/tf-tensorboard",
     visibility = ["//visibility:public"],
     deps = [
-        "//tensorboard/plugins/audio/tf_audio_dashboard",
         "//tensorboard/components/tf_backend",
         "//tensorboard/components/tf_dashboard_common",
-        "//tensorboard/plugins/distributions/tf_distribution_dashboard",
         "//tensorboard/components/tf_globals",
-        "//tensorboard/plugins/graphs/tf_graph_dashboard",
-        "//tensorboard/plugins/histograms/tf_histogram_dashboard",
-        "//tensorboard/plugins/images/tf_image_dashboard",
-        "//tensorboard/components/tf_imports:polymer",
         "//tensorboard/components/tf_imports:lodash",
-        "//tensorboard/plugins/scalars/tf_scalar_dashboard",
+        "//tensorboard/components/tf_imports:polymer",
         "//tensorboard/components/tf_storage",
-        "//tensorboard/plugins/text/tf_text_dashboard",
+        "//tensorboard/plugins/audio/tf_audio_dashboard",
+        "//tensorboard/plugins/distribution/tf_distribution_dashboard",
+        "//tensorboard/plugins/graph/tf_graph_dashboard",
+        "//tensorboard/plugins/histogram/tf_histogram_dashboard",
+        "//tensorboard/plugins/image/tf_image_dashboard",
         "//tensorboard/plugins/projector/vz_projector",
+        "//tensorboard/plugins/scalar/tf_scalar_dashboard",
+        "//tensorboard/plugins/text/tf_text_dashboard",
         "@org_polymer_font_roboto",
         "@org_polymer_iron_icons",
         "@org_polymer_paper_button",
