+++ tensorboard/plugins/graph/tf_graph_app/BUILD
@@ -13,9 +13,9 @@ ts_web_library(
     ],
     path = "/tf-graph-app",
     deps = [
-        "//tensorboard/plugins/graphs/tf_graph_board",
-        "//tensorboard/plugins/graphs/tf_graph_controls",
-        "//tensorboard/plugins/graphs/tf_graph_loader",
+        "//tensorboard/plugins/graph/tf_graph_board",
+        "//tensorboard/plugins/graph/tf_graph_controls",
+        "//tensorboard/plugins/graph/tf_graph_loader",
         "//tensorboard/components/tf_imports:polymer",
         "//tensorboard/components/tf_imports:webcomponentsjs",
         "@org_polymer_iron_component_page",
@@ -27,9 +27,9 @@ tensorboard_webcomponent_library(
     srcs = [":tf_graph_app"],
     destdir = "tf-graph-app",
     deps = [
-        "//tensorboard/plugins/graphs/tf_graph_board:legacy",
-        "//tensorboard/plugins/graphs/tf_graph_controls:legacy",
-        "//tensorboard/plugins/graphs/tf_graph_loader:legacy",
+        "//tensorboard/plugins/graph/tf_graph_board:legacy",
+        "//tensorboard/plugins/graph/tf_graph_controls:legacy",
+        "//tensorboard/plugins/graph/tf_graph_loader:legacy",
         "//third_party/javascript/polymer/v1/iron-component-page:lib",
         "//third_party/javascript/polymer/v1/polymer:lib",
         "//third_party/javascript/polymer/v1/webcomponentsjs:lib",
