+++ tensorboard/plugins/graph/tf_graph_board/BUILD
@@ -10,9 +10,9 @@ ts_web_library(
     srcs = ["tf-graph-board.html"],
     path = "/tf-graph-board",
     deps = [
-        "//tensorboard/plugins/graphs/tf_graph",
-        "//tensorboard/plugins/graphs/tf_graph_common",
-        "//tensorboard/plugins/graphs/tf_graph_info",
+        "//tensorboard/plugins/graph/tf_graph",
+        "//tensorboard/plugins/graph/tf_graph_common",
+        "//tensorboard/plugins/graph/tf_graph_info",
         "//tensorboard/components/tf_imports:polymer",
         "@org_polymer_paper_progress",
     ],
@@ -23,9 +23,9 @@ tensorboard_webcomponent_library(
     srcs = [":tf_graph_board"],
     destdir = "tf-graph-board",
     deps = [
-        "//tensorboard/plugins/graphs/tf_graph:legacy",
-        "//tensorboard/plugins/graphs/tf_graph_common:legacy",
-        "//tensorboard/plugins/graphs/tf_graph_info:legacy",
+        "//tensorboard/plugins/graph/tf_graph:legacy",
+        "//tensorboard/plugins/graph/tf_graph_common:legacy",
+        "//tensorboard/plugins/graph/tf_graph_info:legacy",
         "//third_party/javascript/polymer/v1/paper-progress:lib",
         "//third_party/javascript/polymer/v1/polymer:lib",
     ],
