+++ tensorboard/plugins/graph/tf_graph_dashboard/BUILD
@@ -12,10 +12,10 @@ ts_web_library(
     deps = [
         "//tensorboard/components/tf_backend",
         "//tensorboard/components/tf_dashboard_common",
-        "//tensorboard/plugins/graphs/tf_graph",
-        "//tensorboard/plugins/graphs/tf_graph_board",
-        "//tensorboard/plugins/graphs/tf_graph_controls",
-        "//tensorboard/plugins/graphs/tf_graph_loader",
+        "//tensorboard/plugins/graph/tf_graph",
+        "//tensorboard/plugins/graph/tf_graph_board",
+        "//tensorboard/plugins/graph/tf_graph_controls",
+        "//tensorboard/plugins/graph/tf_graph_loader",
         "//tensorboard/components/tf_imports:polymer",
         "//tensorboard/components/vz_sorting",
     ],
@@ -28,10 +28,10 @@ tensorboard_webcomponent_library(
     deps = [
         "//tensorboard/components/tf_backend:legacy",
         "//tensorboard/components/tf_dashboard_common:legacy",
-        "//tensorboard/plugins/graphs/tf_graph:legacy",
-        "//tensorboard/plugins/graphs/tf_graph_board:legacy",
-        "//tensorboard/plugins/graphs/tf_graph_controls:legacy",
-        "//tensorboard/plugins/graphs/tf_graph_loader:legacy",
+        "//tensorboard/plugins/graph/tf_graph:legacy",
+        "//tensorboard/plugins/graph/tf_graph_board:legacy",
+        "//tensorboard/plugins/graph/tf_graph_controls:legacy",
+        "//tensorboard/plugins/graph/tf_graph_loader:legacy",
         "//tensorboard/components/vz_sorting:legacy",
         "//third_party/javascript/polymer/v1/polymer:lib",
     ],
