+++ tensorboard/plugins/graph/tf_graph_debugger_data_card/BUILD
@@ -14,7 +14,7 @@ ts_web_library(
     deps = [
         "//tensorboard/components/tf_dashboard_common",
         "//tensorboard/components/tf_imports:polymer",
-        "//tensorboard/plugins/graphs/tf_graph_common",
+        "//tensorboard/plugins/graph/tf_graph_common",
         "@org_polymer_paper_material",
         "@org_polymer_paper_slider",
         "@org_polymer_paper_spinner",
@@ -28,7 +28,7 @@ tensorboard_webcomponent_library(
     destdir = "tf-graph-debugger-data-card",
     deps = [
         "//tensorboard/components/tf_dashboard_common:legacy",
-        "//tensorboard/plugins/graphs/tf_graph_common:legacy",
+        "//tensorboard/plugins/graph/tf_graph_common:legacy",
         "//third_party/javascript/polymer/v1/iron-collapse:lib",
         "//third_party/javascript/polymer/v1/iron-list:lib",
         "//third_party/javascript/polymer/v1/paper-icon-button:lib",
