+++ tensorboard/plugins/graph/tf_graph_debugger_data_card/demo/BUILD
@@ -4,13 +4,13 @@ load("//tensorboard/defs:web.bzl", "ts_web_library")
 
 licenses(["notice"])  # Apache 2.0
 
-# bazel run //third_party/tensorflow/tensorboard/plugins/graphs/tf_graph_debugger_data_card/demo
+# bazel run //third_party/tensorflow/tensorboard/plugins/graph/tf_graph_debugger_data_card/demo
 ts_web_library(
     name = "demo",
     srcs = ["index.html"] + glob(["data/**"]),
     path = "/tf-graph-debugger-data-card/demo",
     deps = [
-        "//tensorboard/plugins/graphs/tf_graph_debugger_data_card",
+        "//tensorboard/plugins/graph/tf_graph_debugger_data_card",
         "//tensorboard/components/tf_imports:webcomponentsjs",
         "@org_polymer_iron_demo_helpers",
         "@org_polymer_paper_styles",
