+++ tensorboard/plugins/graph/tf_graph_info/BUILD
@@ -17,8 +17,8 @@ ts_web_library(
     deps = [
         "//tensorboard/components/tf_dashboard_common",
         "//tensorboard/components/tf_imports:polymer",
-        "//tensorboard/plugins/graphs/tf_graph_common",
-        "//tensorboard/plugins/graphs/tf_graph_debugger_data_card",
+        "//tensorboard/plugins/graph/tf_graph_common",
+        "//tensorboard/plugins/graph/tf_graph_debugger_data_card",
         "@org_polymer_iron_collapse",
         "@org_polymer_iron_list",
         "@org_polymer_paper_button",
@@ -36,8 +36,8 @@ tensorboard_webcomponent_library(
     destdir = "tf-graph-info",
     deps = [
         "//tensorboard/components/tf_dashboard_common:legacy",
-        "//tensorboard/plugins/graphs/tf_graph_common:legacy",
-        "//tensorboard/plugins/graphs/tf_graph_debugger_data_card:legacy",
+        "//tensorboard/plugins/graph/tf_graph_common:legacy",
+        "//tensorboard/plugins/graph/tf_graph_debugger_data_card:legacy",
         "//third_party/javascript/polymer/v1/iron-collapse:lib",
         "//third_party/javascript/polymer/v1/iron-list:lib",
         "//third_party/javascript/polymer/v1/paper-icon-button:lib",
