+++ tensorboard/plugins/graph/tf_graph_info/demo/BUILD
@@ -4,15 +4,15 @@ load("//tensorboard/defs:web.bzl", "ts_web_library")
 
 licenses(["notice"])  # Apache 2.0
 
-# bazel run //third_party/tensorflow/tensorboard/plugins/graphs/tf_graph_info/demo
+# bazel run //third_party/tensorflow/tensorboard/plugins/graph/tf_graph_info/demo
 ts_web_library(
     name = "demo",
     srcs = ["index.html"] + glob(["data/**"]),
     path = "/tf-graph-info/demo",
     deps = [
-        "//tensorboard/plugins/graphs/tf_graph_common",
-        "//tensorboard/plugins/graphs/tf_graph_info",
-        "//tensorboard/plugins/graphs/tf_graph_loader",
+        "//tensorboard/plugins/graph/tf_graph_common",
+        "//tensorboard/plugins/graph/tf_graph_info",
+        "//tensorboard/plugins/graph/tf_graph_loader",
         "//tensorboard/components/tf_imports:webcomponentsjs",
         "@org_polymer_iron_demo_helpers",
         "@org_polymer_paper_styles",
