+++ tensorboard/plugins/histogram/tf_histogram_dashboard/BUILD
@@ -22,7 +22,7 @@ ts_web_library(
         "//tensorboard/components/tf_imports:lodash",
         "//tensorboard/components/tf_imports:polymer",
         "//tensorboard/components/tf_runs_selector",
-        "//tensorboard/plugins/histograms/vz_histogram_timeseries",
+        "//tensorboard/plugins/histogram/vz_histogram_timeseries",
         "@org_polymer_iron_collapse",
         "@org_polymer_paper_button",
         "@org_polymer_paper_icon_button",
