+++ tensorboard/plugins/scalar/tf_scalar_dashboard/BUILD
@@ -23,7 +23,7 @@ ts_web_library(
         "//tensorboard/components/tf_paginated_view",
         "//tensorboard/components/tf_runs_selector",
         "//tensorboard/components/tf_storage",
-        "//tensorboard/plugins/scalars/vz_line_chart",
+        "//tensorboard/plugins/scalar/vz_line_chart",
         "@org_polymer_iron_collapse",
         "@org_polymer_paper_button",
         "@org_polymer_paper_checkbox",
